The target master parse language is _no longer_ automatically inserted.  If you enable master parsing, and you specify more than one language, you must use an explicit substitution of $target_lang in the notebook destination; otherwise, bdc will abort. If you only have a single language, you can omit $target_lang.
Added a master.instructor setting, allowing you to control whether or not instructor notebooks are generated. It's true, by default.

Removed the use of a leading "/" in a notebook destination as a means to suppress the automatic insertion of the target language.

Changed the substitution of $extension so that it does not include the leading ".".