Added ability to enable or disable a footer that is automatically added to each generated notebook. The feature is controlled by a per-notebook footer option in the ". (See the sample build.yaml). The feature is on by default, and the default footer is a Databricks copyright.
You can now set master.heading.path and master.footer.path to the string "DEFAULT" to force the internal default to be used, which is useful if overriding a non-default value in the notebook_defaults setting.

Added course_info.copyright_year configuration item, to set the copyright year. Defaults to current year.